design/requirements.txt
"""
requirements for the g-node

1. Progress bar and upload solution
- you need mod_wsgi + apache (configuration as daemon - check multithread is true)
- python-memcached lib has to be installed
- memcached - should be started (memcached -d -m 2048 -M -p 11211, telnet localhost 11211 to check)

2. 
a) Changes in Notification (pinax-source/lib/python2.6/site-packages/notification)
??? Not yet done.. to be completed, Phase II?

b) GRAPELLI: USE THE VERSION WITH CHANGES!
http://code.google.com/p/django-grappelli/source/detail?r=692
(MOVE TO THE MAIN FOLDER)

c) Changes in pinax-source/lib/python2.6/site-packages/wiki:
../lib/python2.6/site-packages/wiki/forms.py
../lib/python2.6/site-packages/wiki/views.py

3. Before going live check whether project files do not contain empty # links!!


"""








--find-links=http://pypi.pinaxproject.com
--find-links=http://pypi2.pinaxproject.com

# Get all requirements formerly known as "external_libs" explicitly
# hashlib
python-dateutil==1.4.1
docutils==0.5
elementtree==1.2.7-20070827-preview

gdata==1.3.3
Pygments==1.0
python-openid==2.2.4
python-twitter==0.5
python-yadis==1.1.0
simplejson==2.0.9
textile==2.1.3
vobject==0.8.1c
BeautifulSoup==3.0.7a
ybrowserauth==1.2
Markdown==2.0.1
diff-match-patch==20090804
geopy==0.93dev-r0
pytz==2009l
creole==1.0.1

django-ajax-validation==0.1.3
django-announcements==0.1.0
django-atomformat==0.1.0dev
django-avatar==1.0.2
django-bookmarks==0.1.0
django-db-log==1.0.1
django-debug-toolbar==0.7.0
django-email-confirmation==0.1.3
django-extensions==0.4.1
django-flag==0.1.1
django-friends==0.1.4
django-gravatar==0.1.0
django-locations==0.1.1
django-mailer==0.1.0
django-messages==0.4.2
django-microblogging==0.1.2
django-notification==0.1.4
django-oembed==0.1.0
django-openid==0.2.0
django-pagination==1.0.5.1
django-robots==0.6.1
django-swaps==0.5.3
django-sorting>=0.1
django-tagging==0.3
django-threadedcomments==0.5.1
django-timezones==0.1.4
django-uni-form==0.6.0
django-voting==0.1_pre
django-wikiapp==0.2.0
django-photologue==2.3
django-attachments==0.3
django-markup==0.3
template_utils==0.4p2
django-staticfiles==0.1.2
